+++ Domibus-MSH/src/main/java/eu/domibus/core/pmode/provider/PModeProvider.java
@@ -122,7 +122,7 @@ public abstract class PModeProvider {
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public void removePMode(int id) {
+    public void removePMode(long id) {
         LOG.debug("Removing PMode with id: [{}]", id);
         configurationRawDAO.deleteById(id);
     }
